requirements.txt
cryptography==43.0.3
Flask==3.0.3
GitPython==3.1.43
igdb-api-v4==0.3.3
libgravatar==1.0.4
mistletoe==1.4.0
praw==7.8.1
py-cord==2.6.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
